inference/requirements.txt
pyyaml
tqdm
pandas
wandb
langchain
langchain-openai
pydantic
fastapi
